Phase K: ServerRedirect long→int32_t + README polish
PacketStructures.h: ServerRedirect's sector_id and ip_address are now
explicitly int32_t instead of long, so the wire layout is identical on
Linux (sizeof(long)=8) and Win32 (sizeof(long)=4). Payload is now 10
bytes everywhere with port at offset 8.

master_join_test.cpp: collapse the historical {8, 16} offset probe pair
to a single offset-8 assertion now that the wire size is consistent.

README: punctuation cleanup (em-dash → hyphen/semicolon), acknowledge
that the Net-7 current codebase is private/inaccessible so this is the
best fork to build on for now, trim the redundant "what this means
practically" paragraph (precedence rules already covered above).

+++ proxy/PacketStructures.h
@@ -3,6 +3,8 @@
 #ifndef _PACKET_STRUCTURES_H_INCLUDED_
 #define _PACKET_STRUCTURES_H_INCLUDED_
 
+#include <stdint.h>
+
 struct EnbTcpHeader
 {
     short   size;
@@ -254,8 +256,8 @@ struct GlobalTicket
 
 struct ServerRedirect
 {
-    long    sector_id;
-    long    ip_address;
+    int32_t sector_id;     // explicitly 4-byte on all platforms (Phase K)
+    int32_t ip_address;    // — Win32 client expects 4-byte fields on the wire
     short   port;
 } ATTRIB_PACKED;
